Replace older SchemaWriterFactory with new JacksonServiceResponseSchema.
Note that the older SchemaWriterFactory has already been just renamed to JacksonServiceResponseSchema.

1. Split to |ServiceResponseSchema| and |JacksonServiceResponseSchema|.
It allows to accept other formats than JSON by implementing a new class
such as |JAXBServiceResponseSchema|.

2. Remove naming configuration with |TaskSource| (e.g. "column_name"),
and introduce |ValueLocator|. It allows to locate values in the response
not just by top-level attribute names but by JSONPath or jq for example.

3. Introduce an interface |ServiceResponseSchemaBuildable| to implement
a factory of |ServiceResponseSchema|, instead of implementing a method
|buildSchemaWriterFactory| in |RestClientInputPluginBase|.